Added tpe_db database scripts. Added projection classes. Added another base entity class. Refactored the entity classes, struts.xml, web.xml.

+++ tpe/src/main/java/org/cgiar/dapa/ccafs/tpe/dao/IClimateDao.java
@@ -25,7 +25,7 @@ import org.cgiar.dapa.ccafs.tpe.entity.Station;
  * @author NMATOVU
  *
  */
-public interface IClimateDao extends IGenericDao<Climate, Integer> {
+public interface IClimateDao extends IGenericDao<Climate, Long> {
 
 	/**
 	 * Retrieves the climate records for the specified category from the
